requirements.txt
beautifulsoup4
mega.py
moviepy
nekobin
Pillow
py-deezer
pyaes
pyrogram
pyshorteners
python-dotenv
python-telegram-bot
pytube
pytz
qrcode
requests
selenium
six
tornado
tzlocal
urllib3
wget
youtube-dl
ffmpeg
requests_toolbelt
